test(e2e): Generate list of pages to test using Eleventy
This switches from using a hard-coded list of pages in the end-to-end tests to a list automatically generated by Eleventy.

It's not ideal as it means that the file has to exist before `e2e/constants.ts` is valid, but it works (for `npm start` too).

I also tried to use a dynamic import, but it means you can't parametrise the tests in the normal way (as it needs an `await`). (There is `test.step()`, but that affect parallelism and other things.)

Eleventy also has `--to=json` CLI option (which spits out the entire site as JSON), but it doesn't help much and it's a bit unwieldy as it includes all the page content.