ci: Add workflow to create GitHub releases on merge
This adds a GitHub Actions workflow to automatically create a GitHub release when a PR is merged to main.

This is a simplified version of the workflow in https://github.com/defencedigital/moduk-frontend. Publishing to npm is not required here. It also does not bother updating the version `package.json` as the package isn't being published. (The version in package.json has been changed to make this clear.)

As previously, the approach was tested in https://github.com/jpveooys/github-actions-testing-v2.