feat(layout): Add phase banner
This adds a beta phase banner across the entire site. It links to our email address.

I had to place it inside `<header>` to avoid https://dequeuniversity.com/rules/axe/4.1/region. This seems to be the way GOV.UK is going as well: https://github.com/alphagov/govuk-frontend/pull/2824#issuecomment-1387248436.

There was also some complexities with styling as the bottom border of the phase banner should not be shown when the primary nav is visible. A `moduk--primary-nav-expanded` class is used on `<body>` for this (dynamically added and removed on mobile when the menu is opened and closed).