ci: Fix error pushing tag during release
This resolves a problem where a release would fail if two PRs were merged in quick succession.

The failure was due to `git push --follow-tags` failing on the release build for the first PR due it the commit being behind the latest `main`.

This instead makes release-it push only the relevant tag. It was tested in https://github.com/jpveooys/github-actions-testing-v2.