feat(home): Update Updates section
This updates the Updates section on the home page in advance of v1.0.0 of `@moduk/frontend` being released. Have put an approximate date in.

The link goes to the main releases page on GitHub (not a specific release) as it's not likely to be feasible to update the site on every insignificant release.