test(visual-regression): Skip JavaScript disabled for component pages
We're not getting much value from the visual regression tests with JavaScript disabled on the component pages, as the pages are quite long with all the code snippets expanded.

It also makes more sense to focus the JavaScript disabled visual regression tests on the components specific to this repo, which are covered by the tests on other pages.