deps: Update various dependencies
There were quite a few out-of-date dependencies, so this picks off most of the easy updates. No updates to dependencies with breaking changes have been made.

This includes a Playwright update. WebKit now appears to support the clipboard in Linux so there were some changes to the code snippet tests. There was also a problem with the toMatch assertion which required a @ts-expect-error: https://github.com/microsoft/playwright/issues/21453