test(visual-regression): Use overlay file system when running tests
This updates `scripts/run-playwright-container.sh` so that an overlay file system is used when mounting the current directory (including node_modules). This means `node_modules` starts off the same as the host, but any changes written to it in the container are temporary and limited to the container itself.

The `e2e` directory is mounted normally so that reports and screenshots are written to the host.

This speeds up the `npm install` part of `npm run test:visual` for me (on Ubuntu).