build: Fix component previews build on Windows
Component previews weren't working on Windows (with `glob` v8 or v9) because the Windows path separator, `\`, is an escape character in globs.

This enables `windowsPathsNoEscape` in all globs as the simplest fix. This changes the meaning of `\` so it works on Windows as a path separator, but can't be used to escape special characters.