feat(error-pages): Add page not found page
This adds a page not found page that displays for 404 errors.

This required providing a custom nginx.conf to implement properly, as the default one contained an empty location block for `/404.html` (which should be set to internal so `/404.html` isn't a valid page itself).

Using a custom `nginx.conf` is in line with examples of how to use the container image anyway (e.g. https://github.com/sclorg/nginx-container/tree/1864ed11f9b2062f3ffb97e1651842775933ab1e/examples).

The logging configuration was merged into the new `nginx.conf` as part of this. This includes configuring `error_log` in the root context (instead of `http`), which captures more events.

Finally, the default description was removed from the base layout as it was out of date and the other examples of a 404 I checked did not have a description.