feat(colours): Add page
This adds the colours page at `/styles/colours/`. No links to the page are added as part of this change.

There was some complexity in extracting the hex values of the colours out of Sass. The CSS `contents` property is not accessible, so that can't be used. There are some libraries for extracting variables out of Sass, but the they all seemed to have problems and limitations. Therefore I ended up with the approach of using Sass to output the colour values as CSS variables, and parsing the hex codes out of that.