test(visual-regression): Use smaller tablet device
This switches the `tablet-safari` project in Playwright from the 'iPad (gen 7)' device to 'iPad (gen 6)'.

'iPad (gen 7)' was falling into the desktop breakpoint and so not actually triggering the tablet layout. 'iPad (gen 6)' is a bit smaller so does trigger the tablet layout.

The difference can be seen, for example, in the header on the home page.

Tags were also brought in line with https://github.com/defencedigital/moduk-frontend/pull/213.