ci: Build and test container image
This adds a new job to the Lint and test GitHub Actions workflow that:

- builds the container image
- starts a container using the image
- makes sure the root URL doesn't return an error, and contains the text 'Design and build consistent Defence services'

I also investigated making Playwright run against the container on GitHub Actions (instead of using `npx serve`). This proved complicated and cumbersome, due to there being multiple jobs needing the container image, and the Playwright jobs already running in a container (making it complex to start a container within the job). It does not seem worth it for the added complexity to the workflow.