feat(split-layout): Add back to top link
This adds a back to top link on all pages that use the split layout.

On tablet and desktop, when JavaScript is available, the back to top link is hidden if the entire side nav fits in the viewport.

Also on on tablet and desktop, the back to top link becomes fixed in the bottom-left corner when scrolling through a long page.

On mobile, the link appears near the end of the page.

The link is not contained within a landmark, in line with https://design-system.service.gov.uk/. This triggers an axe failure, so the link is excluded from axe. (Putting it in its own `<nav>` was also an option, but it wasn't clear that that'd be useful – screen readers often have a keyboard shortcut to scroll to the top of the page already.)