test(visual-regression): Change full-page screenshot logic
While working on #200, I had some problems with full-page screenshots in Firefox. Because we were using the document element to take the screenshot, it was scrolling to the middle of the page before taking the screenshot, causing some unexpected results with the floating back-to-top link.

This switches to taking full-page screenshots using the `page` object instead. I removed the screenshot-splitting logic for now as it's no longer being used anywhere (we can always re-add it if we need it again),