feat(examples): Add React code snippets
This adds a React tab to examples for supported components containing a code snippet.

The code for the snippet is imported from the `@moduk/frontend` package and run through Prettier to reduce the line length.

If a matching React example is not found, the React tab is hidden.

The code snippets are only shown if the SHOW_REACT_CODE_SNIPPETS environment variable is set to true. I haven't added parallel test suites on CI as that doesn't seem worth the effort here, but some tests for the React code snippet run if that environment variable is set.